requirements.txt
django>=5.0
djangorestframework
django-environ
pillow
python-dotenv
openai
